refactor(rbac): collapse Can* helpers into single can.go
Phase 2 PR 3 of the Server consolidation.

- Merged instance_can.go + space_can.go into a single cli/internal/core/can.go
  organized by scope (server-level, space-scoped, room-scoped). The "space"
  helpers keep their existing names through the dual-RBAC transition; renames
  to Server* happen in PR 10 alongside the schema flip.
- Dropped the transitional shim helpers introduced in PR 1: CanSpaceList,
  CanSpaceJoin, CanSpaceCreate, CanAdminSpacesView, CanAdminSpaceDelete,
  CanListSpace, CanJoinSpace. Their underlying permissions were dropped per
  ADR-028 and the helpers were standing in only to keep callers compiling.
- Updated callers (graph resolvers + a couple of tests) to inline the
  appropriate behavior:
    - Viewer.canCreateSpace / canListSpaces: always true.
    - Viewer.canAdminViewSpaces: delegates to CanAdminAccess.
    - Space.viewerCanJoinSpace: gates on auth only.
    - Mutation.createSpace / joinSpace: drop the permission gate (auth still
      required); per ADR-028 these become no-op gates and the mutations
      themselves go away in PR 10.
    - Query.spaces: drop the per-user space.list filter; discovery is
      unrestricted post-consolidation.
- Test files updated to drop tests that exercised the removed helpers.

Verified:
- mise x -- go build ./...: green
- mise x -- go test ./...: green (modulo pre-existing TestAuthRoutes_TestEmailEndpoint)
- frontend build / svelte-check unaffected (no schema changes)

Part of #284.
