--- cli/internal/core/instance_can.go
@@ -1,123 +1,0 @@
-package core
-
-import "context"
-
-// This file provides semantic helper functions for instance-level permission checks.
-// These wrap the low-level HasInstancePermission function with business-meaningful names,
-// making code more readable and permission usage easier to audit.
-//
-// Each function returns (bool, error) where:
-//   - bool indicates whether the user has the permission
-//   - error is non-nil only if there was a system error checking permissions
-//
-// Note: These functions check RBAC permissions only. Config-based admin check
-// (owners.emails) should be done separately by the caller.
-
-// CanSpaceList checks if a user can view the list of spaces.
-//
-// Transitional shim: the underlying space.list permission is dropped per ADR-028.
-// During the dual-RBAC transition (PR 1–9) this helper stays for resolver
-// compatibility and always returns true — every user can see the (single)
-// server. The helper itself is removed in PR 3.
-func (c *ChattoCore) CanSpaceList(ctx context.Context, userID string) (bool, error) {
-	return true, nil
-}
-
-// CanSpaceJoin checks if a user can join spaces.
-//
-// Transitional shim: the underlying space.join permission is dropped per ADR-028.
-// Joining the server is registering an account; not gated separately.
-// The helper itself is removed in PR 3.
-func (c *ChattoCore) CanSpaceJoin(ctx context.Context, userID string) (bool, error) {
-	return true, nil
-}
-
-// CanSpaceCreate checks if a user can create new spaces.
-//
-// Transitional shim: the underlying space.create permission is dropped per ADR-028
-// (no nested servers). Returning true preserves the existing dev workflow during
-// PRs 1–9 while spaces still exist as a transitional concept; the helper and its
-// callers go away in PR 10.
-func (c *ChattoCore) CanSpaceCreate(ctx context.Context, userID string) (bool, error) {
-	return true, nil
-}
-
-// CanAdminAccess checks if a user can access the admin panel.
-// Only instance admins have this permission.
-func (c *ChattoCore) CanAdminAccess(ctx context.Context, userID string) (bool, error) {
-	return c.HasInstancePermission(ctx, userID, PermAdminAccess)
-}
-
-// CanAdminUsersView checks if a user can view the users page in admin.
-func (c *ChattoCore) CanAdminUsersView(ctx context.Context, userID string) (bool, error) {
-	return c.HasInstancePermission(ctx, userID, PermAdminUsersView)
-}
-
-// CanAdminUsersManage checks if a user can edit user role assignments.
-func (c *ChattoCore) CanAdminUsersManage(ctx context.Context, userID string) (bool, error) {
-	return c.HasInstancePermission(ctx, userID, PermAdminUsersManage)
-}
-
-// CanAdminSpacesView checks if a user can view the spaces page in admin.
-//
-// Transitional shim: the underlying admin.view-spaces permission is dropped
-// per ADR-028 (no spaces admin page post-consolidation). For now we delegate to
-// CanAdminAccess to keep the existing admin nav working while spaces still
-// exist; the helper goes away in PR 10 along with the spaces admin page.
-func (c *ChattoCore) CanAdminSpacesView(ctx context.Context, userID string) (bool, error) {
-	return c.CanAdminAccess(ctx, userID)
-}
-
-// CanAdminRolesView checks if a user can view the roles page in admin.
-func (c *ChattoCore) CanAdminRolesView(ctx context.Context, userID string) (bool, error) {
-	return c.HasInstancePermission(ctx, userID, PermAdminRolesView)
-}
-
-// CanAdminRolesManage checks if a user can create/edit instance roles and their permissions.
-func (c *ChattoCore) CanAdminRolesManage(ctx context.Context, userID string) (bool, error) {
-	return c.HasInstancePermission(ctx, userID, PermAdminRolesManage)
-}
-
-// CanAdminSystemView checks if a user can view the system and data pages in admin.
-func (c *ChattoCore) CanAdminSystemView(ctx context.Context, userID string) (bool, error) {
-	return c.HasInstancePermission(ctx, userID, PermAdminSystemView)
-}
-
-// CanDMView checks if a user can access the DM space and read DMs.
-// Verified users have this permission by default.
-func (c *ChattoCore) CanDMView(ctx context.Context, userID string) (bool, error) {
-	return c.HasInstancePermission(ctx, userID, PermDMView)
-}
-
-// CanDMWrite checks if a user can start DM conversations and send messages.
-// Verified users have this permission by default.
-func (c *ChattoCore) CanDMWrite(ctx context.Context, userID string) (bool, error) {
-	return c.HasInstancePermission(ctx, userID, PermDMWrite)
-}
-
-// CanAdminManageUser checks if an actor can perform admin user-management
-// actions (e.g. editing identity, clearing cooldowns) on a target user based
-// on instance role hierarchy. Self-management is always allowed; otherwise
-// the actor's highest role must outrank the target's highest role.
-//
-// Note: this checks RBAC hierarchy only. Config-based admins (owners.emails)
-// are not visible to the RBAC engine and should bypass this check at the
-// resolver layer (they outrank everyone).
-func (c *ChattoCore) CanAdminManageUser(ctx context.Context, actorID, targetID string) (bool, error) {
-	if actorID == targetID {
-		return true, nil
-	}
-	return c.instanceRBACEngine.CanUserManageUser(ctx, actorID, targetID)
-}
-
-// CanDeleteUser checks if an actor can delete a specific user account.
-// Returns true if:
-//   - The actor is deleting their own account and has user.delete.self permission, OR
-//   - The actor has the user.delete permission (admin capability)
-func (c *ChattoCore) CanDeleteUser(ctx context.Context, actorID, targetUserID string) (bool, error) {
-	if actorID == targetUserID {
-		return c.HasInstancePermission(ctx, actorID, PermUserDeleteSelf)
-	}
-
-	return c.HasInstancePermission(ctx, actorID, PermUserDelete)
-}
